Move asset input organism
