+++ packages/app/src/features/easy-borrow/components/form/Borrow.tsx
@@ -2,7 +2,7 @@ import { TokenWithBalance } from '@/domain/common/types'
 import { TokenSymbol } from '@/domain/types/TokenSymbol'
 import { IconButton } from '@/ui/atoms/icon-button/IconButton'
 import { Panel } from '@/ui/atoms/panel/Panel'
-import { AssetInput } from '@/ui/organisms/new/asset-input/AssetInput'
+import { AssetInput } from '@/ui/organisms/asset-input/AssetInput'
 import { cn } from '@/ui/utils/style'
 import { testIds } from '@/ui/utils/testIds'
 import { raise } from '@/utils/assert'
