+++ packages/app/src/features/easy-borrow/components/form/Deposits.tsx
@@ -5,7 +5,7 @@ import { Button, ButtonIcon } from '@/ui/atoms/button/Button'
 import { Link } from '@/ui/atoms/link/Link'
 import { Panel } from '@/ui/atoms/panel/Panel'
 import { Info } from '@/ui/molecules/info/Info'
-import { AssetInput } from '@/ui/organisms/new/asset-input/AssetInput'
+import { AssetInput } from '@/ui/organisms/asset-input/AssetInput'
 import { cn } from '@/ui/utils/style'
 import { testIds } from '@/ui/utils/testIds'
 import { NormalizedUnitNumber } from '@marsfoundation/common-universal'
